Fix Unbeaten paths challenge data initialization

+++ src/DataStructuresAlgorithms.h
@@ -349,7 +349,7 @@ vector<string> FindItinerary(vector<vector<string>> &, string const &);
 void EulerianPath(string, map<string, multiset<string>> &, vector<string> &);
 size_t EggDrops(size_t, size_t);
 vector<string> fizzBuzz(size_t);
-vector<size_t> UnbeatenPaths(size_t, vector<vector<size_t>> &, size_t);
+vector<long> UnbeatenPaths(size_t, vector<vector<size_t>> &, size_t);
 string RoadsInHackerland(size_t, vector<vector<size_t>> &);
 string RoadsInHackerland1(size_t, vector<vector<size_t>> &);
 string RoadsInHackerland2(size_t, vector<vector<size_t>> &);
